requirements.txt
PyAutoGUI
PyGetWindow
opencv-python-headless
numpy
pandas
pywin32
pyyaml
tqdm
matplotlib
seaborn
requests
ipython
psutil
dxcam
onnxruntime_directml
obs-websocket-py
cupy-cuda11x
obws_python
simpleobsws>=1.3.0